Update package metadata, build system, and version info (#2)
This commit makes more updates to the package metadata to remove
unnecessary double quotes from some lines. It also modernizes the build
configuration a bit by using PBR as a PEP517 builder. It also adds the
missing version string from PBR (which adds a runtime dependency on it
the library).